Media section in the assessment page

+++ apps/backend/database/migrations/2021_11_23_162628_add_document_media_columns.php
@@ -15,7 +15,7 @@ class AddDocumentMediaColumns extends Migration
     {
         Schema::table('documents', function (Blueprint $table) {
             $table->tinyInteger('position')->unsigned()->default(0)->comment('Optional position of item');
-            $table->string('description')->nullable()->comment('Optional description of document');
+            $table->string('comments')->nullable()->comment('Optional description of document');
             $table->json('exif_data')->nullable()->comment('json exif data for media files');
         });
     }
@@ -28,9 +28,15 @@ class AddDocumentMediaColumns extends Migration
     public function down()
     {
         Schema::table('documents', function (Blueprint $table) {
-            $table->dropColumn('position');
-            $table->dropColumn('description');
-            $table->dropColumn('exif_data');
+            if (Schema::hasColumn('documents', 'position')) {
+                $table->dropColumn('position');
+            }
+            if (Schema::hasColumn('documents', 'comments')) {
+                $table->dropColumn('comments');
+            }
+            if (Schema::hasColumn('documents', 'exif_data')) {
+                $table->dropColumn('exif_data');
+            }
         });
     }
 }
